+++ src/apps/lwaftr/fragmentv4_test.lua
@@ -3,7 +3,6 @@ if type((...)) == "string" then module(..., package.seeall) end
 
 local constants = require("apps.lwaftr.constants")
 local fragmentv4 = require("apps.lwaftr.fragmentv4")
-local fragv4_h = require("apps.lwaftr.fragmentv4_hardened")
 local eth_proto = require("lib.protocol.ethernet")
 local ip4_proto = require("lib.protocol.ipv4")
 local lwutil = require("apps.lwaftr.lwutil")
@@ -222,230 +221,12 @@ function test_dont_fragment_flag()
 end
 
 
-local pattern_fill, pattern_check = (function ()
-   local pattern = { 0xCC, 0xAA, 0xFF, 0xEE, 0xBB, 0x11, 0xDD }
-   local function fill(array, length)
-      for i = 0, length-1 do
-         array[i] = pattern[(i % #pattern) + 1]
-      end
-   end
-   local function check(array, length)
-      for i = 0, length-1 do
-         assert(array[i], pattern[(i % #pattern) + 1], "pos: " .. i)
-      end
-   end
-   return fill, check
-end)()
-
-
-function test_reassemble_pattern_fragments()
-   print("test:   length=1046 mtu=520 + reassembly")
-
-   local pkt = make_ipv4_packet(1046 - ip4_proto:sizeof() - eth_proto:sizeof())
-   pattern_fill(pkt.data + ip4_proto:sizeof() + eth_proto:sizeof(),
-                pkt.length - ip4_proto:sizeof() - eth_proto:sizeof())
-
-   local code, result = fragmentv4.fragment(pkt, 520)
-   assert(code == fragmentv4.FRAGMENT_OK)
-   assert(#result == 3)
-
-   assert(pkt_payload_size(result[1]) + pkt_payload_size(result[2]) +
-          pkt_payload_size(result[3]) == 1046 - ip4_proto:sizeof() - eth_proto:sizeof())
-
-   local size = pkt_payload_size(result[1]) + pkt_payload_size(result[2]) + pkt_payload_size(result[3])
-   local data = ffi.new("uint8_t[?]", size)
-
-   for i = 1, #result do
-      local ih = get_ihl_from_offset(result[i], constants.ethernet_header_size)
-      ffi.copy(data + pkt_frag_offset(result[i]),
-               result[i].data + eth_proto:sizeof() + ih,
-               pkt_payload_size(result[i]))
-   end
-   pattern_check(data, size)
-end
-
-
-function test_reassemble_two_missing_fragments(vlan_id)
-   print("test:   two fragments (one missing)")
-   local pkt = assert(make_ipv4_packet(1200), vlan_id)
-   local code, fragments = fragmentv4.fragment(pkt, 1000)
-   assert(code == fragmentv4.FRAGMENT_OK)
-   assert(#fragments == 2)
-
-   local frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-end
-
-
-function test_reassemble_three_missing_fragments(vlan_id)
-   print("test:   three fragments (one/two missing)")
-   local pkt = assert(make_ipv4_packet(1000))
-   local code, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   assert(code == fragmentv4.FRAGMENT_OK)
-   assert(#fragments == 3)
-
-   local frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[3])))
-
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[3])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[3])))
-
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[3])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[1])))
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   frag_table = fragv4_h.initialize_frag_table(20, 5)
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[2])))
-   assert(fragv4_h.FRAGMENT_MISSING ==
-          (fragv4_h.cache_fragment(frag_table, fragments[3])))
-end
-
-
-function test_reassemble_two(vlan_id)
-   print("test:   payload=1200 mtu=1000")
-   local pkt = assert(make_ipv4_packet(1200), vlan_id)
-   assert(pkt.length > 1200, "packet shorter than payload size")
-
-   -- Keep a copy of the packet, for comparisons
-   local orig_pkt = packet.clone(pkt)
-   -- A single error above this can lead to packets being on the freelist twice...
-   assert(pkt ~= orig_pkt, "packets must be different")
-
-   local code, fragments = fragmentv4.fragment(packet.clone(orig_pkt), 1000)
-   assert(code == fragmentv4.FRAGMENT_OK)
-   assert(#fragments == 2)
-   assert(fragments[1].length ~= 0, "fragment[1] length must not be 0")
-   assert(fragments[2].length ~= 0, "fragment[2] length must not be 0")
-
-   local function try(f)
-      local frag_table = fragv4_h.initialize_frag_table(20, 5)
-      local code, pkt
-      for i=1,#f do
-         code, pkt = fragv4_h.cache_fragment(frag_table, f[i])
-      end
-      assert(code == fragv4_h.REASSEMBLY_OK, "returned: " .. code)
-      assert(pkt.length == orig_pkt.length)
-
-      for i = 1, pkt.length do
-         if i ~= 24 and i ~= 25 then
-            assert(pkt.data[i] == orig_pkt.data[i],
-                   "byte["..i.."] expected="..orig_pkt.data[i].." got="..pkt.data[i])
-         end
-      end
-   end
-
-   try { fragments[1], fragments[2] }
-   _, fragments = fragmentv4.fragment(packet.clone(orig_pkt), 1000)
-   assert(fragments[1].length ~= 0, "fragment[1] length must not be 0")
-   assert(fragments[2].length ~= 0, "fragment[2] length must not be 0")
-   try { fragments[2], fragments[1] }
-end
-
-
-function test_reassemble_three(vlan_id)
-   print("test:   payload=1000 mtu=400")
-   local pkt = assert(make_ipv4_packet(1000), vlan_id)
-
-   -- Keep a copy of the packet, for comparisons
-   local orig_pkt = packet.clone(pkt)
-
-   local code, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   assert(code == fragmentv4.FRAGMENT_OK)
-   assert(#fragments == 3)
-   assert(orig_pkt.length == 1034, "wtf")
-
-   local function try(f)
-      local frag_table = fragv4_h.initialize_frag_table(20, 5)
-      local code, pkt
-      for i=1,#f do
-         code, pkt = fragv4_h.cache_fragment(frag_table, f[i])
-      end
-      assert(code == fragv4_h.REASSEMBLY_OK, "returned: " .. code)
-      assert(pkt.length == orig_pkt.length)
-
-      for i = 1, pkt.length do
-         if i ~= 24 and i ~= 25 then
-            assert(pkt.data[i] == orig_pkt.data[i],
-                   "byte["..i.."] expected="..orig_pkt.data[i].." got="..pkt.data[i])
-         end
-      end
-   end
-
-   try { fragments[1], fragments[2], fragments[3] }
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   try { fragments[2], fragments[3], fragments[1] }
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   try { fragments[3], fragments[1], fragments[2] }
-
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   try { fragments[3], fragments[2], fragments[1] }
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   try { fragments[2], fragments[1], fragments[3] }
-   _, fragments = fragmentv4.fragment(packet.clone(pkt), 400)
-   try { fragments[1], fragments[3], fragments[2] }
-end
-
-
 function selftest()
    print("test: lwaftr.fragmentv4.fragment_ipv4")
    test_payload_1200_mtu_1500()
    test_payload_1200_mtu_1000()
    test_payload_1200_mtu_400()
    test_dont_fragment_flag()
-   test_reassemble_pattern_fragments()
-
-   local function testall(vlan_id)
-      local suffix = " (no vlan tag)"
-      if vlan_id then
-         suffix = " (vlan id=" .. vlan_id .. ")"
-      end
-      print("test: lwaftr.fragv4_h.cache_fragment_ipv4" .. suffix)
-      test_reassemble_two_missing_fragments(vlan_id)
-      test_reassemble_three_missing_fragments(vlan_id)
-      test_reassemble_two(vlan_id)
-      test_reassemble_three(vlan_id)
-   end
-   testall(nil)
-   testall(42)
 end
 
 -- Run tests when being invoked as a script from the command line.
