+++ src/program/lwaftr/setup.lua
@@ -13,6 +13,7 @@ local lwutil     = require("apps.lwaftr.lwutil")
 local basic_apps = require("apps.basic.basic_apps")
 local pcap       = require("apps.pcap.pcap")
 local ipv4_apps  = require("apps.lwaftr.ipv4_apps")
+local ipv4_reassemble = require("apps.ipv4.reassemble")
 local arp        = require("apps.ipv4.arp")
 local ipv6_apps  = require("apps.lwaftr.ipv6_apps")
 local ndp        = require("apps.lwaftr.ndp")
@@ -99,7 +100,7 @@ function lwaftr_app(c, conf, device)
    local iexternal_interface = queue.external_interface
    local iinternal_interface = queue.internal_interface
 
-   config.app(c, "reassemblerv4", ipv4_apps.Reassembler,
+   config.app(c, "reassemblerv4", ipv4_reassemble.Reassembler,
               { max_ipv4_reassembly_packets =
                    gexternal_interface.reassembly.max_packets,
                 max_fragments_per_reassembly_packet =
